requirements.txt
Package              Version
-------------------- -----------
anyio                3.6.2
argon2-cffi          21.3.0
argon2-cffi-bindings 21.2.0
asttokens            2.1.0
attrs                22.1.0
backcall             0.2.0
beautifulsoup4       4.11.1
bleach               5.0.1
certifi              2022.9.24
cffi                 1.15.1
charset-normalizer   2.1.1
click                8.1.3
debugpy              1.6.3
decorator            5.1.1
defusedxml           0.7.1
entrypoints          0.4
executing            1.2.0
fastjsonschema       2.16.2
idna                 3.4
importlib-metadata   5.0.0
importlib-resources  5.10.0
ipykernel            6.17.1
ipython              8.6.0
ipython-genutils     0.2.0
jedi                 0.18.1
Jinja2               3.1.2
joblib               1.2.0
jsonschema           4.17.0
jupyter_client       7.4.5
jupyter_core         5.0.0
jupyter-server       1.23.1
jupyterlab-pygments  0.2.2
kafka-python         2.0.2
MarkupSafe           2.1.1
matplotlib-inline    0.1.6
mistune              2.0.4
nbclassic            0.4.8
nbclient             0.7.0
nbconvert            7.2.4
nbformat             5.7.0
nest-asyncio         1.5.6
nltk                 3.7
notebook_shim        0.2.2
numpy                1.23.4
oauthlib             3.2.2
packaging            21.3
pandocfilters        1.5.0
parso                0.8.3
pexpect              4.8.0
pickleshare          0.7.5
pip                  22.0.4
pkgutil_resolve_name 1.3.10
platformdirs         2.5.3
prometheus-client    0.15.0
prompt-toolkit       3.0.32
psutil               5.9.4
ptyprocess           0.7.0
pure-eval            0.2.2
py4j                 0.10.9.5
pycparser            2.21
Pygments             2.13.0
pyparsing            3.0.9
pyrsistent           0.19.2
pyspark              3.3.1
python-dateutil      2.8.2
pyzmq                24.0.1
regex                2022.10.31
requests             2.28.1
requests-oauthlib    1.3.1
Send2Trash           1.8.0
setuptools           56.0.0
six                  1.16.0
sniffio              1.3.0
soupsieve            2.3.2.post1
stack-data           0.6.0
terminado            0.17.0
textblob             0.17.1
tinycss2             1.2.1
tornado              6.2
tqdm                 4.64.1
traitlets            5.5.0
tweepy               4.12.1
urllib3              1.26.12
wcwidth              0.2.5
webencodings         0.5.1
websocket-client     1.4.2
zipp                 3.10.0
